requirements.txt
aiofiles>=23.2.1
aiohttp>=3.12.0
aiosignal>=1.3.2
annotated-types>=0.7.0
anyio>=4.9.0
arxiv>=2.0.0
attrs>=25.3.0
beautifulsoup4>=4.12.2
bs4>=0.0.1
certifi>=2025.4.26
click>=8.2.1
colorama>=0.4.6
# datetime is part of Python standard library
fastapi>=0.104.1
importlib-metadata>=8.7.0
jinja2>=3.1.2
json-repair>=0.29.8
json5>=0.9.25
langchain_community>=0.3.17
langchain-openai
langchain-core>=0.3.60
langchain-text-splitters>=0.3.8
langgraph>=0.2.73,<0.3
loguru>=0.7.3
lxml>=5.4.0
markdown>=3.8
markdown2>=2.5.3
mistune>=3.1.3
numpy>=2.2.6
openai>=1.82.0
# pathlib is part of Python standard library
pydantic>=2.11.5
pymupdf>=1.26.0
python-docx>=1.1.2
python-dotenv>=1.1.0
pyyaml>=6.0.2
regex>=2024.11.6
requests>=2.32.3
starlette>=0.46.2
tiktoken>=0.9.0
tqdm>=4.67.1
typing-extensions>=4.13.2
urllib3>=2.4.0
uvicorn>=0.34.2
# uuid is part of Python standard library
websockets>=15.0.1
md2pdf>=1.0.1
ddgs
docx2pdf
python-multipart
htmldocx

# Azure App Service specific packages
gunicorn>=21.2.0
azure-functions>=1.18.0